Move classes from outside lib/ to PSR-4
