+++ apps/files_sharing/appinfo/app.php
@@ -26,15 +26,13 @@
  *
  */
 
-namespace OCA\Files_Sharing\Appinfo;
-
 $l = \OC::$server->getL10N('files_sharing');
 
 \OC::$CLASSPATH['OC_Share_Backend_File'] = 'files_sharing/lib/share/file.php';
 \OC::$CLASSPATH['OC_Share_Backend_Folder'] = 'files_sharing/lib/share/folder.php';
 \OC::$CLASSPATH['OC\Files\Storage\Shared'] = 'files_sharing/lib/sharedstorage.php';
 
-$application = new Application();
+$application = new \OCA\Files_Sharing\AppInfo\Application();
 $application->registerMountProviders();
 
 \OCA\Files_Sharing\Helper::registerHooks();
