+++ apps/files_sharing/lib/API/OCSShareWrapper.php
