+++ apps/files_sharing/lib/AppInfo/Application.php
